requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
xlsxwriter>=3.2
python-pptx>=0.6.21
pytest>=8.0
